types package has been moved from vm dir to smartcontract
