+++ testcase/smartcontract/neovm/hash/hash160.go
@@ -1,13 +1,14 @@
 package hash
 
 import (
+	"math/big"
+	"time"
+
 	sdkcom "github.com/ontio/ontology-go-sdk/common"
 	"github.com/ontio/ontology-go-sdk/utils"
 	"github.com/ontio/ontology-test/testframework"
+	"github.com/ontio/ontology/smartcontract/types"
 	"github.com/ontio/ontology/vm/neovm"
-	"github.com/ontio/ontology/vm/types"
-	"math/big"
-	"time"
 )
 
 func TestHash160(ctx *testframework.TestFrameworkContext) bool {
