+++ testcase/smartcontract/neovm/hash/hash256.go
@@ -1,13 +1,14 @@
 package hash
 
 import (
+	"math/big"
+	"time"
+
 	sdkcom "github.com/ontio/ontology-go-sdk/common"
 	"github.com/ontio/ontology-go-sdk/utils"
 	"github.com/ontio/ontology-test/testframework"
+	"github.com/ontio/ontology/smartcontract/types"
 	"github.com/ontio/ontology/vm/neovm"
-	"github.com/ontio/ontology/vm/types"
-	"math/big"
-	"time"
 )
 
 func TestHash256(ctx *testframework.TestFrameworkContext) bool {
@@ -57,7 +58,7 @@ func TestHash256(ctx *testframework.TestFrameworkContext) bool {
 	return true
 }
 
-func hash256(input []byte) []byte{
+func hash256(input []byte) []byte {
 	return new(neovm.ECDsaCrypto).Hash256(input)
 }
