+++ testcase/smartcontract/neovm/utils/asbytearray_string.go
@@ -1,12 +1,13 @@
 package utils
 
 import (
+	"math/big"
+	"time"
+
 	sdkcom "github.com/ontio/ontology-go-sdk/common"
 	"github.com/ontio/ontology-go-sdk/utils"
 	"github.com/ontio/ontology-test/testframework"
-	"github.com/ontio/ontology/vm/types"
-	"math/big"
-	"time"
+	"github.com/ontio/ontology/smartcontract/types"
 )
 
 func TestAsByteArrayString(ctx *testframework.TestFrameworkContext) bool {
