+++ testcase/smartcontract/neovm/utils/take.go
@@ -1,13 +1,14 @@
 package utils
 
 import (
+	"math/big"
+	"time"
+
 	sdkcom "github.com/ontio/ontology-go-sdk/common"
 	"github.com/ontio/ontology-go-sdk/utils"
 	"github.com/ontio/ontology-test/testframework"
 	"github.com/ontio/ontology/common"
-	"github.com/ontio/ontology/vm/types"
-	"math/big"
-	"time"
+	"github.com/ontio/ontology/smartcontract/types"
 )
 
 func TestTake(ctx *testframework.TestFrameworkContext) bool {
