+++ testcase/smartcontract/testcase.go
@@ -18,12 +18,14 @@
 package smartcontract
 
 import (
+	"github.com/ontio/ontology-test/testcase/smartcontract/native"
+	"github.com/ontio/ontology-test/testcase/smartcontract/neovm"
 	"github.com/ontio/ontology-test/testcase/smartcontract/wasmvm"
 )
 
 //Register test case
 func TestSmartContract() {
-	//native.TestNative()
-	//neovm.TestNeoVM()
+	native.TestNative()
+	neovm.TestNeoVM()
 	wasmvm.TestWasmVM()
 }
